backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
boto3==1.36.4
python-magic==0.4.27
Pillow==11.1.0
WeasyPrint==63.1
markdown==3.7
pdfplumber==0.11.4
slowapi==0.1.9
pytest==8.3.4
httpx==0.28.1
moto[s3]==5.0.27
pytest-asyncio==0.25.0
python-docx==1.1.2
